# Optimize some queries that get tags or validations