Resolves #2474 

Improves load time of some web pages by optimizing queries that get label tags.

I tested on my local dev environment, running an up to date Seattle database dump (our largest database atm). Speed improvements will be smaller for smaller databases.
* Loading the admin page went from 45 seconds to 20 seconds
* Loading an admin version of a user's page went from 8 seconds to 2 seconds for a user with a lot of labels. If the user has very few labels, the load time should be around 2 seconds regardless.
* Loading the validation page had no significant change in load time, but I updated the query for labels to validate in the same way. This added consistency and allowed me to delete a few helper functions we were using for the old method.

The reason we see speedups in the first two cases is due to the large number of labels being queries; max 15k for admin, max 1k for admin user dashboard, but less than 100 for the validation page.

@ThatOneGoat @kumararoosh I also tried this approach for Gallery, but it actually slowed things down a little bit. The reason is that the way you have the queries right now, you're querying for _all_ labels that match the description, then you only grab the tags for the labels that you're going to use. If instead we only queries for a number of labels on the first go, and wrapped the whole thing in a loop so we could search again if we don't find enough with non-expired imagery, we might see a minor performance improvement. However, I don't think the added complexity is worth the trouble, given how minor the impact would be (the bigger time sink is verifying that the panos exist).