@jonfroehlich 2021-03-05
This is great! What do you mean by “ The reason we see speedups in the
first two cases is due to the large number of labels being queries; max 15k
for admin, max 1k for admin user dashboard”

The part that my brain cannot parse is the wording of a “large number of
labels being queries”

On Thu, Mar 4, 2021 at 11:09 PM Mikey Saugstad <notifications@github.com>
wrote:

> Merged #2481
> <https://github.com/ProjectSidewalk/SidewalkWebpage/pull/2481> into
> develop.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/ProjectSidewalk/SidewalkWebpage/pull/2481#event-4411978968>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAML55LHG4FXKSUF2ERFA5DTCB7SRANCNFSM4YUVZ2WA>
> .
>
-- 
Sent from phone