@misaugstad 2021-03-05
Well I guess part of that is a typo: "labels being queries" -> "labels being queried".

The admin page gets the tags for all of the labels in the "contributions" tab (up to 15k labels), whereas we only get the set of tags for a small number of labels when loading the validation page.