@jonfroehlich 2021-03-06
Aha, thanks!

On Fri, Mar 5, 2021 at 3:09 PM Mikey Saugstad <notifications@github.com>
wrote:

> Well I guess part of that is a typo: "labels being queries" -> "labels
> being queried".
>
> The admin page gets the tags for all of the labels in the "contributions"
> tab (up to 15k labels), whereas we only get the set of tags for a small
> number of labels when loading the validation page.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/ProjectSidewalk/SidewalkWebpage/pull/2481#issuecomment-791780710>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAML55OHY4LJ6C3RNT3DONLTCFQC5ANCNFSM4YUVZ2WA>
> .
>


-- 
Jon E. Froehlich <https://jonfroehlich.github.io/> (he/him
<https://www.mypronouns.org/he-him>) | @jonfroehlich
<https://twitter.com/jonfroehlich>
Associate Professor, Allen School of Computer Science & Engineering
Core Faculty, Urban Design and Planning, College of Built Environments
Director, Makeability Lab <https://makeabilitylab.cs.washington.edu/> |
Associate Director, CREATE <http://create.uw.edu/>
University of Washington
Help make sidewalks more accessible: http://projectsidewalk.io